move and rename core state file
