+++ src/app/items/conatainers/items-container/items-container.component.spec.ts
@@ -4,7 +4,7 @@ import {ItemsContainerComponent} from './items-container.component';
 import {Store} from '@ngrx/store';
 import {ItemsComponent} from '../../components/items/items.component';
 import {MockStore, TestingModule} from '../../../../testing/utils';
-import {AppState} from '../../../reducers';
+import {AppState} from '../../../core/core.state';
 import {CoreModule} from '../../../core/core.module';
 
 type ItemsFeatureStateSlice = Pick<AppState, 'items'>;
